+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 1.2.1 - Unreleased
 
 - Fix scrolling issues (double scrollbar and scrolling down on load) that occur in Trilium v0.60+.
+- Fix the `$dateCreated` and `$dateModified` properties displaying an empty value in Trilium v0.61+.
+  - For Trilium v0.61.0 to v0.61.5, `$dateCreated` is not available and `$dateModified` will display the modification date of the note's content and not the note itself.
+  - If you are using the v0.61 beta, upgrade to at least v0.61.6 for these dates to work as intended.
 - Fix the `header` attribute setting displaying the attribute name instead of an empty header cell when set to an empty value (`#attribute=name,header=`).
 
 ## 1.2.0 - 2023-01-22